+++ docs/regression_curator/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this gear are documented in this file.
 
+## To be released
+
+* Adds `debug` argument and sets verbose logging statement level to debug
+
 ## 0.1.1
 
 * Changes type to `analysis`